Store property_definitions.position as NUMERIC, not REAL
Addresses PR feedback: REAL is single-precision float, so drag-reorder
midpoint inserts (`new.position = (prev.position + next.position) / 2`)
accumulate rounding error and will eventually collide on adjacent
values, forcing a rebalance. NUMERIC is exact-precision — the standard
choice for fractional ordering keys.

NUMERIC(20, 10) gives 10 decimal places of scale, well beyond what any
realistic subdivision sequence needs. Column uses `asdecimal=False` on
the Python side so downstream serializers keep seeing a float and no
callers have to deal with Decimal.

Migration revision id unchanged (20260422_0074). Apply on existing
dev / test DBs with:

    alembic downgrade 20260421_0073 && alembic upgrade head

+++ backend/alembic/versions/20260422_0074_add_property_definitions.py
@@ -201,7 +201,12 @@ def upgrade() -> None:
             name VARCHAR(100) NOT NULL,
             type property_type NOT NULL,
             applies_to property_applies_to NOT NULL DEFAULT 'both',
-            position REAL NOT NULL DEFAULT 0,
+            -- NUMERIC (not REAL) so drag-reorder midpoint inserts — which set
+            -- the new row's position to the fractional mean of its neighbors —
+            -- never hit float-precision rounding and force a rebalance.
+            -- (20,10) gives 10 decimal places of scale: plenty of headroom
+            -- for subdivisions while staying compact.
+            position NUMERIC(20, 10) NOT NULL DEFAULT 0,
             color VARCHAR(7),
             options JSONB,
             created_at TIMESTAMPTZ NOT NULL,
